[#321] Move check provider logic into the actual celery worker for sending notification. Fix unrelated linting and talisman errors
