+++ .talismanrc
@@ -85,3 +85,5 @@ fileignoreconfig:
       checksum: 8e45e9081aabcec8ce66555af127449d0abd78b4faa1260a95a8c3c4563e2e6a
     - filename: app/models.py
       checksum: cd38660ab8d5d7672ad5bd141f93d1737ad967e260fafa613df844a2f3225fb1
+    - filename: app/delivery/send_to_providers.py
+      checksum: 231849229bc5e142d9e28e1f34338f63178d1f7ea357f2589acb84816055bdb7
